requirements.txt
pandas
numpy
librosa
resampy
scikit-learn
matplotlib
tqdm
tensorflow
python-box
pyyaml
ensure